Rework the way the map is drawn.  Drawing individual quads is too cpu and gpu intensive.  Instead, use SpriteBatch to draw to a render target using BackToFront.  We specify a unique layer depth for each object asset.  Then by using BackToFront it means that tiles are effectively grouped by texture and thus texture swapping is minimal.  Then we render this single texture to a single large quad which represents the map in 3D space.
If this does not fix the performance issues we can try one more thing - convert the asset textures into a large single texture.  This should not be necessary.

By rendering to one large quad we do remove the possibility of being able to make tile-by-tile 3D effects.  We probably don't need them.  But if we do then we can instead draw individual quads and use UV texture indexes instead.